ADEN-9566 Fix ad-engine load on search pages
