+++ app/services/ads/ads.js
@@ -56,6 +56,7 @@ export default Service.extend({
   },
 
   setupAdsContext(adsContext) {
+    console.error('SETUP ADS CONTEXT!');
     adsContext.user = {
       isAuthenticated: this.currentUser.isAuthenticated,
     };
